Add udev rules for creating links for nvme devices
As per the document at [1], EBS volumes on Nitro instances are exposed as NVME
devices, and do not have consistent device numbering. This commit adds the
mechanism used by Amazon Linux 2 to the Ubuntu AMI in order to create links for
each device to the mapped ID (e.g. `/dev/xsdf`) based on the NVME device ID.

Unfortunately this currently uses Python, so that needs installing too - this
should be rectified soon.

[1]: https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/nvme-ebs-volumes.html