Make build time part of AMI name
Packer does not yet support defaults in variables, so it's not possible
to get a consistent timestamp to use as part of the AMI and a tag. Since
AMI name must be unique, it's better to put the build timestamp there
rather than in a tag.

This bug is tracked upstream as hashicorp/packer#9430.